Add more labels to issues
This change adds a hard coded list of candidate labels that can be taken
from the labels field of a ticket and added as a label to the generated
gh issue.

The list is hard coded because so many tickets have duplicate or
borderline meaningless labels that you wouldn't want to add to a
gh issue. So we hard code the list of candidate labels.